// src/services/schemas.ts
export const CreateAuthorSchema = z.object({
  firstName: z.string().describe("The first name of the author"),
  lastName: z.string().describe("The last name of the author"),
  birthDate: z.string().optional().describe("The birth date of the author in ISO format"),
  biography: z.string().optional().describe("Biography of the author"),
  nationality: z.string().optional().describe("Nationality of the author"),
});

export const UpdateAuthorSchema = z.object({
  firstName: z.string().optional().describe("The first name of the author"),
  lastName: z.string().optional().describe("The last name of the author"),
  birthDate: z.string().optional().describe("The birth date of the author in ISO format"),
  biography: z.string().optional().describe("Biography of the author"),
  nationality: z.string().optional().describe("Nationality of the author"),
});

export const GetAuthorSchema = z.object({
  authorId: z.number().describe("The unique numeric ID of the author. MUST be extracted from the user request."),
});

export const SearchAuthorsSchema = z.object({
  query: z.string().describe("Search query for author name"),
});